dev(retention): rewrote gatherer and added many filter controls (#328)
I rewrote our Gatherer from scratch. This is the service that converts a filter into a set of customers (as well as generates stats). Typical filter usage is the default settings plus bucket, and restrict vehicles. With those values, my new gatherer (Super Gatherer) reduces CPU usage by 94% and is around 87% faster, and also reduces the amount of time MySQL is computing by 84% which will lighten the load on our rds server as well as our web server.

In addition to this, I've added multiple new filter controls, as these all required handlers for the new gatherer, so I figured it's the best time to implement them.

https://vicimus.atlassian.net/browse/BUMP-6440
https://vicimus.atlassian.net/browse/BUMP-8652
https://vicimus.atlassian.net/browse/BUMP-8252
https://vicimus.atlassian.net/browse/BUMP-6478
https://vicimus.atlassian.net/browse/BUMP-6410
https://vicimus.atlassian.net/browse/BUMP-6502
https://vicimus.atlassian.net/browse/BUMP-6503
https://vicimus.atlassian.net/browse/BUMP-7806