dev(shared): make the tran() method way more effecient (#367)
It exists in two places, so i just kept them the same.

If the app's locale is english, we don't have to translate anything. If a default is provided, then that is just the english word, we can just turn around and return the english word and save all the look ups. This eliminates 99% of all overhead for translating.
If it can't do the above, it will cache the results now so it only has to do the lookup once.
https://vicimus.atlassian.net/browse/GBX-5936
https://vicimus.atlassian.net/browse/GBX-5933
https://vicimus.atlassian.net/browse/GBX-5932