dev(shared): improve the compiler to be awesome (#405)
Caching can happen outside of this class, so removed the two
dependencies that were required just for caching. Made it able to scan
and include sub-files within the directory automatically, so we no
longer need to explicitly include them. Added support for the complex
translation types that contain context and file locations.
https://vicimus.atlassian.net/browse/BUMP-12309