dev(shared): merge php74 into master  (#409)
Whenever we develop new features for Marketingsuite, we develop these
new features, interfaces, services, etc in the php74 branch. These
changes and new developments are never merged into master, but if we
eventually upgrade Marketingsuite to php8, we will be switching to use
the master branch of support, and we need all these changes to be
present or a bunch of work and fixes will just disappear and things will
be broken.